Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main(){
-    float d;
-    scanf("%.4f",&d);
+    double d;
+    scanf("%lf",&d);
     printf("You entered: %.4f",d);
     return 0;
 }
